Common: Move file_utils.py, secure_directory.py, secure_file.py to common/utils/file_utils/
