+++ monkey/common/utils/file_utils/secure_file.py
@@ -4,14 +4,14 @@ import stat
 from contextlib import contextmanager
 from typing import Generator
 
-from .operating_system import is_windows_os
+from ..operating_system import is_windows_os
 
 if is_windows_os():
     import win32file
     import win32job
     import win32security
 
-    from . import windows_permissions
+    from .. import windows_permissions
 
 
 logger = logging.getLogger(__name__)
